refactor: move vis type map out of component
It's needed in other places.
