+++ packages/favorites-dialog/src/EnhancedTable.js
@@ -26,24 +26,9 @@ import YearOverYearLineIcon from './icons/YearOverYearLineIcon';
 import YearOverYearColumnIcon from './icons/YearOverYearColumnIcon';
 import SingleValueIcon from './icons/SingleValueIcon';
 import PivotTableIcon from './icons/PivotTableIcon';
+import visTypeMap from './visTypes';
 
 import { changePage, setRowsPerPage, sortData, selectFavorite } from './actions';
-
-const visTypeMap = {
-    BAR: { label: i18n.t('Bar'), icon: BarIcon },
-    STACKED_BAR: { label: i18n.t('Stacked bar'), icon: StackedBarIcon },
-    COLUMN: { label: i18n.t('Column'), icon: ColumnIcon },
-    STACKED_COLUMN: { label: i18n.t('Stacked column'), icon: StackedColumnIcon },
-    LINE: { label: i18n.t('Line'), icon: LineIcon },
-    AREA: { label: i18n.t('Area'), icon: AreaIcon },
-    PIE: { label: i18n.t('Pie'), icon: PieIcon },
-    RADAR: { label: i18n.t('Radar'), icon: RadarIcon },
-    GAUGE: { label: i18n.t('Gauge'), icon: GaugeIcon },
-    YEAR_OVER_YEAR_LINE: { label: i18n.t('Year over year (line)'), icon: YearOverYearLineIcon },
-    YEAR_OVER_YEAR_COLUMN: { label: i18n.t('Year over year (column)'), icon: YearOverYearColumnIcon },
-    SINGLE_VALUE: { label: i18n.t('Single value'), icon: SingleValueIcon },
-    PIVOT_TABLE: { label: i18n.t('Pivot table'), icon: PivotTableIcon },
-};
 
 const Time = ({ date }) => {
     const d = new Date(date);
